*: rename Headlamp component to Web UI
The reasoning is that this component is the main UI for Lokomotive
clusters and we'll probably add more functionality than what is in
upstream Headlamp.
