--- assets/charts/components/headlamp/templates/service.yaml
@@ -1,15 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "headlamp.fullname" . }}
-  labels:
-    {{- include "headlamp.labels" . | nindent 4 }}
-spec:
-  type: {{ .Values.service.type }}
-  ports:
-    - port: {{ .Values.service.port }}
-      targetPort: {{ .Values.service.targetPort }}
-      protocol: TCP
-      name: http
-  selector:
-    {{- include "headlamp.selectorLabels" . | nindent 4 }}
