+++ cli/cmd/component.go
@@ -26,7 +26,6 @@ import (
 	_ "github.com/kinvolk/lokomotive/pkg/components/external-dns"
 	_ "github.com/kinvolk/lokomotive/pkg/components/flatcar-linux-update-operator"
 	_ "github.com/kinvolk/lokomotive/pkg/components/gangway"
-	_ "github.com/kinvolk/lokomotive/pkg/components/headlamp"
 	_ "github.com/kinvolk/lokomotive/pkg/components/httpbin"
 	_ "github.com/kinvolk/lokomotive/pkg/components/istio-operator"
 	_ "github.com/kinvolk/lokomotive/pkg/components/linkerd"
@@ -38,6 +37,7 @@ import (
 	_ "github.com/kinvolk/lokomotive/pkg/components/rook"
 	_ "github.com/kinvolk/lokomotive/pkg/components/rook-ceph"
 	_ "github.com/kinvolk/lokomotive/pkg/components/velero"
+	_ "github.com/kinvolk/lokomotive/pkg/components/web-ui"
 )
 
 var componentCmd = &cobra.Command{
