+++ test/ingress/web-ui/ingress_test.go
@@ -1,0 +1,28 @@
+// Copyright 2020 The Lokomotive Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build aws aws_edge
+// +build e2e
+
+package webui_test
+
+import (
+	"testing"
+
+	testingress "github.com/kinvolk/lokomotive/test/ingress"
+)
+
+func TestWebUIIngress(t *testing.T) {
+	testingress.AccessIngress(t, "lokomotive-system", "web-ui", "")
+}
